Update jquery.reveal.js
Added a config option to allow you to suppress the modal window from opening when first instantiated. This is particularly useful when opening the modal pragmatically.  You can set the config options once (when the document is ready) and reuse the modal by simply calling .reveal() without defining the options again.